Refactor dev-flags -> auto-solve + clean up code a bit
Not sure why I had called this `dev-flags` before. But... it's an
auto-Game solver. So renaming it to `auto-solve`.

Also, extract magic values into constants.

Also, remove the increased delay when there is only 1 cell remaining.
We no longer need that since we're now showing Just-Ended Games for an
extended duration, even on refresh. i.e. there's no risk of a race
condition causing the game-ending move to error and cause the page to
reload and then suddenly be on the New Game view.