Cache Just-ended Game content for 3 minutes
The motivation here was to make it so that the Just-ended game message
was consistent for all players and across all page refreshes. e.g. if
the game-ending message was `Argh! 😧 Whhhyyyyyy?!` then it should be
that for everyone and even if they refresh.

Re 3 minutes: Because we only show the Just-ended Game for 3 minutes.
Refreshes beyond that will show the New Game view.

Note:
For now, at least, we exclude `games/just_ended/title` from the cache
fragment so that refreshes can update the `T+MM:SS` timer. However,
we may be able to adjust that later after we have a JS timer
incrementing this--depending on implementation.