Make Home page - New Game title stay same for 30s at a time
While I've chosen 30s as a happy medium for caching, the primary goal
here is to make the message sync across all players/refreshes. So they
get the same, but otherwise randomized, message to start.

This follows in the footsteps of our previous commit which did the
same for the just-ended Game message.