"Animate" the T+MM:SS timestamp for just-ended Games
To support showing a "live" timestamp for just-ended Games, we utilize
our existing elapsed_time_controller.js. But, we must extend it as this
is a slightly different use case. Now:

```
There are 2 distinct modes of operation:
1. Formatted: Always match the provided format: `HH:MM:SS`, `MM:SS`, or
   `SS`.
2. Compact: Only display as much as is needed from `HH`, `MM`, and `SS`.
The default is "Compact" mode.
 -> To switch to "Formatted" mode, supply a `formatValue` String.
```

This allows us to always show the minutes so that our T+ timestamp is
consistently shown as `T+MM:SS` (and then stops at a max value of
`T+59:59+`).

I considered changing the browser tooltip to also dynamical reveal the
actual, current T+MM:SS value, but I don't think it's better. It doesn't
change in real time while hovering, so it also adds a little confusion
as the timestamp will quickly be out of date.

Note: Previously, using `format: { type: String }` wasn't working as
expected... Stimulus was reporting true for `this.hasFormatValue` when
one wasn't being supplied. But changing to `format: String` fixed that
issue.