Make the ElapsedTime JS timers more accurate/reliable
Previously, we were passing in an elapsed time in seconds as the start
value. This was inherently less reliable as, reloading the page or
navigating away and coming back would skip updates and even return to
the last known value. I'm not sure how I didn't notice this for so
long...

To fix: The start value we pass in now is, essentially, a start
timestamp, though it's expressed as an Unix Timestamp (integer). From
this, we can accurately determine the elapsed time from the start vale
even if we refresh the page or navigate away for a while and come back.
This is because we can easily calculate the elapsed time from a start
"date".

We also now start by updating the timer on connect. This eliminates the
long second of "incorrectness" when navigating away/back. i.e. if one
navigates away at 00:05, waits 5 seconds, and then navigates back, it
was taking a full second before the timer would change from what was in
the saved turbo data: `00:05` to the correct value: `00:10`. Now, that
update happens virtually instantly instead.