Add Games#spam; use to exclude spam from Index / Tallies
Spam games are games that were just not a valid attempt at completing
the game. There's no automatic detection of this, it's based on feel...
but won't be applied liberally either. This comes out of a case where a
good friend was hammering the game as fast as they could just for fun--
without regard to the gravity of the ongoing battle between the Alliance
and the Mines! So, as the dev, I can now go in and mark these Games as
spam on the Rails console.
- Spam games will still show up on the User's profile page, though, as
  if there's nothing wrong them.
- They can also still be visited via direct URL, etc. They just don't
  show up in the main Games Index page, is all. (This could change in
  the future, of course.)

Example usage:
- First, hand-pick a Game or range of Games to mark as spam, then:

```
Game.for_id(197).console.mark_as_spam

Game.for_id(199..240).update_all(spam: true)
```