Show total Games count next to Win/Loss Tallies
Instead of forcing users to do the math in their head, we can show the
total count of Games included in the tallies. e.g.:

Instead of just
`Alliance: 62 / Mines: 276`

We will no show:
`Alliance: 62 / Mines: 276 Σ 338`

Care is taken to prevent word-wrapping, or, when it is needed, to do it
in opportune places.