Tidy: Refactor / cleanup Collapse Container
This effort was guided by the realizations that:
1. What I've called a "Collapse Container" is known as a
   "Details/Summary" element in HTML 5. And, while native browser
   support does exist for `details` elements, it's actually not good
   enough to use (I tried)... because we can't get CSS transitions out
   of them when in the collapsed state. Resources:
   - https://developer.mozilla.org/en-US/docs/Web/HTML/Element/details
   - https://stackoverflow.com/questions/38213329/how-to-add-css3-transition-with-html5-details-summary-tag-reveal
2. However, we *can( use `group-` pseudo selectors to style the button
   and collapse container by emulating the "open" state of the `details`
   element. i.e. just putting an `open` class in the collapse container.

Realizing number 2 is possible allows us to remove logic from
collapse_controller.js, keeping things simpler via just straight
Tailwind CSS selectors. Cool!

Diving into this inspired further refactoring to keep this widget as
simple as possible for future reuse/applications.

Also, not sure why `Games::Current::Rules::CollapseContainer::Context`
was even in there as it didn't add anything. Plus, I have since moved
away from the pattern of defining the Context for a thing *inside* of
that thing. I now consider that an anti-pattern. Instead, the Context
should always be defined *outside* of the thing it serves.

Finally, we now make the "Rules of Engagement" collapse contain button
appear as interactive when in the open state (not when collapsed).
Hopefully this helps people to realize it's a collapsible section... as
my interaction tracking shows no uses of this feature yet in production.