Fix: Active link underline wasn't showing for Account Page
The `overflow-hidden` class in the `ul` (container) element also causes
the bottom border for active usernames to be hidden. This happens
because, technically, the `li` elements are overflowing the height of
the `ul` container simply by having top/bottom padding.

This somewhat unexpected feature, is why we are using `ul>li` for these
nav links in the first place... so I didn't want to try to change
that.

Instead, just switching from `overflow-hidden` to `overflow-x-clip`
fixes the issue with the red bottom border being hidden. I hadn't
previously understood how `overflow-clip` differed from
`overflow-hidden`... but I guess this is a notable use case difference
(even if I still don't understand why).