Show Time Zone form on User Account page (as well)
We already show this form on the Games Index page, but it makes sense
to add it to the User Account show page as well. This may actually be
more used since it'll be more visible.

The one on the Games Index page is cooler, though, because, on-change,
the page reloads and you can see the Games jump around between days
based on the current Time Zone. For that reason, I'm not consolidating
to just the one on the Account show page, but, rather, keeping it in
both places.

Note: I considered trying to prevent the broadcasted page refresh from
occurring on the User Account Show page, but decided against worrying
with it. It's probably a good thing (for any future changes, especially)
to ensure that the current page is up-to-date--as far as Time Zones go.