Make Fleet Roster open all enough to fit content as needed
If one's name is really long, it's chopped off handily by the Roster
slide menu. Trying out this expand-on-hover idea to see if it's good
in practice.

Also, move the "Roster" heading to the left a bit, to line up with the
right edge of the usernames (instead of the right edge of the
statusmojis). I think this look better.