Show "local time" on User Profile page
Alongside the user's Time Zone, it's handy to be able to see their
local time. While it's not animated (yet?), that could come later...

Note: We add an exception-based "no value indicator" alternative, for
displaying both the Time Zone and this new "local time" indicator. This
should cover a wider range of conditions where the User's Time Zone may
be invalid.

Also, fix up the Account Show and Account Profile page view/styling.
We need to use a table for uniform first column widths, but then we
need to treat the table rows like flex-cols on mobile. This solution
seems to do that...