Fix white-space removal between inline text elements
In a few cases, where we have white-space inserted by the structure of
the HTML, I have historically just added a negative margin to one of
the elements to eliminate the unwanted white space. However, this is a
bit ... 'fudgy'. And, I actually noticed that it wasn't rendered
properly on the Tesla browser (in-car).

So this commit removes these 2 known instances and, instead of fudging
away the white-space, we now utilize `inline-flex` to make it just
cease to exist.