Add games.active_participants_count counter cache
Add Game#active_participants_count and games.active_participants_count
as a counter_cache for this so that we can easily display how many
users actively participated in a Game. This is especially useful to
prevent N+1 queries... as I'd like to include this count in the
Sweep Ops Archive (see next commit).

Also, because I asserted to Grok that ParticipantsTransactions will
never be deleted, it suggested adding
`dependent: :restrict_with_exception` to Game's
`has_many :participant_transactions` association, which is a great idea.