Display Active Participants Count in Game listings/statuses
Previously:
- Game Statuses showed an Alliance win as `⛴️🎉`

Now:
- Game Statuses will show an Alliance win as just `🎉`

And we will now use the `⛴️` emoji just to indicate the number of
active participants in the Game. With e.g. `⛴️ x2`, etc.
- Also, we'll only show the active participants count on the Game
  listings in the Sweep Ops Archive for games with many (>1) active
  participants. This way, we can more easily extract this fact by
  scanning. (Less information to parse.)

These updates apply to all of:
- Game listings in Sweep Ops Archive
- Game listings in User Profile page
- Just-ended Game Show -> Title / Status
- Past Game Show -> Title / Status
However, the User -> Game Show pages don't include Game status right
now. I think because I didn't want to show it for the "Bests" games,
but it carried over to the listings.

Note: To support the fine-tuned spacing I wanted arounder the
participant count indicators, we change the game listing links to not
use an underline, but rather, to use a thing bottom border. This
prevents breaks in the underline where `gap-x-*` or `mx-*` styles are
used.