Use pencil SVG over Emoji + Extract `.icon-link` CSS class
First, we switch form using a `✏️` emoji to using a heroicon SVG. This
allows for more consistency across Operating Systems. (Especially
Tesla's browser.) Plus it allows for more consistent link styles. Which
means showing the underline effect on the text, but not the icon. The
new standard here being the "Authentication Link" and "Voluntary
Discharge" link on the Account Show page.

To support this, we extract an `icon-link` CSS class, which does the
necessary `underline` work, while preventing a `border-b-*` style.
(Preventing border-b will be more important in the next commit, where
we switch link styles to always use a border-b instead of an underline.)

Since we're no longer using `Emoji.pencil` in any core application code,
we remove this method definition as well.

Also: Style the "Authentication Link" in green and the "Voluntary
Discharge" link in red. This should add a bit more of an "intuitive"
understanding of the operations being performed.