Use `border-b-*` instead of `underline` for link decoration
The problem with `underline` is that it presents a fragmented line any
time we add things like `margin` or `gap` spacing between link elements.
Borders don't have this problem. (In fact, we utilize the fragmented
line to our advantage via the new `icon-link` CSS class that we added
in the previous commit.)

In order to return the link styling to normalcy after this change, we:
- Remove use of `nav` elements outside of header/footer so that we
  aren't forced into a `border-b-0` style.
- Add `.active-nav` (and replace `.no-active-links` `.inactive-nav`) for
  controlling when to use "active" links outside of (and inside of)
  `nav` elements.
  - The main need here was to be able to have "active" links (links that
    maintain a border-b style when active) outside of `nav` elements.
    The rename of `.no-active-links` -> `.inactive-nav` was for
    parallelism.
- Utilize `.btn-basic` for ensuring no `border-b` style on e.g. the
  Cancel (`🙈`) buttons.