Separate out the Users->Bests->Games display case view
Previously, we had an issue where the Game Status wasn't showing for
Users past games, regardless of context. i.e.
- Both for User "Bests" Games (shown in the display case), and
- Past User Games (not show in a display case)
... but we should only be excluding the Game Status when displayed in
the display case. Because only then is it already obvious what the
Game Status was (all "Bests" are victories).

To fix, we separate out the Users Bests Games from the Users Games
views.

We also take care, on the display-case version, to just return a
`no_content` header if the User or Game can't be found during the
request (should never really happen). This avoids a useless re-rendering
of the Home page that will never be consumed since the expected turbo
frame won't be present there.

Finally, I've made a slight general coding style tweak on `require_*`
methods, so I carried that over to the PatternsController and included
that with this commit since it's the only other relevant code to change.