Add apple-touch-icon-precomposed.png
Just adding this to silence the errors I see in the console when copying
a link into this app. The error comes from the Paste app (macOS)
requesting details about the link so it can compose a nice link preview
for it. For some reason, it also wants this file
`apple-touch-icon-precomposed.png`, though I don't see anywhere it's
used based on the link preview after.

The `apple-touch-icon-precomposed.png` file is fairly common, though, so
I don't mind adding it. By "precomposed" we mean, the non-transparent
version of the `apple-touch-icon.png` image. However, our
`apple-touch-icon.png` image does not have a transparent background, so
this simply duplicates that file with the new name.

Previously, the error from the server was:
```
Started GET "/apple-touch-icon-precomposed.png" for 127.0.0.1 at 2025-03-09 16:11:23 -0500

ActionController::RoutingError (No route matches [GET] "/apple-touch-icon-precomposed.png"):
```