Indicate which user ended the game as 🎉 or 💣
Show which user ended the game (and how) from the Duty Roster for
Just-ended / Past Games. The idea here is to make it clear which user
caused the Game-ending move, which is especially useful when there is
more than one active participant.

This is already data we were collecting, so just needed to update the
view to render the indicators.