Fix up wording in Roster Tooltip
Just trying to be a bit more explicit/clear.

Also: Still can't figure out how to make the tooltip not have to sort
of "animate in" on first load, but... pretty sure it has (at least
mostly) to do with floating-ui trying to calculate the position to move
the tooltip into view on open. Setting the position to bottom-end
doesn't help either, and is an uglier result since the arrow pops off
the end of the tooltip a bit--another issue I never fully figured out.