Don't display Fleet Mustering notification for old Games
I've found that, in prod, if the conditions are right we can end up
seeing a "Fleet Mustering!" notification for a quickly-failed new Game.

This was because *not* delivering the Fleet Mustering notification was
entirely dependent on the Just-ended Game Footer partial for Active
Participants not being present. i.e. if the Game Over view for an
active participant was present, then the Fleet Mustering notification
would be shown. This meant that all it took was a delayed background
job and a quickly-ended Game to cause a false positive on showing
the "Fleet Mustering" notification.

To fix, we scope the "Fleet Mustering" notification to a turbo stream
channel named after the previous (just-ended) Game. This way, the only
way to receive "Fleet Mustering" notifications is to be on the same
Just-ended Game screen for which a new Game was just created.