Abstract "previous" and "next" game finder logic
Previously, I think we had essentially business logic in the View
Models. Either way, it makes sense to define a service object for this.

We use just a module instead of a class because there's no need to
instantiate an object here.

By injecting the `base_arel` (default: `Game`) we allow for additional
filtering as needed. e.g. for User-scoped Games.

This was extracted after seeing the pattern, which will be needed in
one additional spot in the following commit.