Show the Roster tooltip on the left i/o bottom
This prevents the "animate-in" effect we were seeing before. Which, now
seems fairly apparent: was caused by floating-ui.js trying to push the
tooltip into the viewable area of the screen. Loading-in on the left to
start doesn't have this problem.

I'm not sure I like the left-oriented tootltip better, visually, but it
does keep the participation status icons in view, which is nice.