Fix fickleness w/ tooltip anchor hover targets
Previously, the Roster tooltip anchor, in particular, felt quite fickle.
I think this one, especially because its anchor element contains an SVG.
Though I don't know why that affects it exactly.

To fix: Add `pointer-events-none` on the anchor element

At the same time, though, it felt like the surface area for mouseover of
the anchor target was not as large as the SVG it contained... so I've
also added an additional padding/buffer around the anchor element for
mouseovers/clicks. This plays out fine for other elements so far as
I've tested, by design: since the negative margin that exactly
matches/inverts the positive padding.