Don't pause the auto-solver when Cell Highlights present
The auto-solver--app/javascript/controllers/auto_solve_controller.js--
knows to skip a beat if/when it's going too fast. It detects this by
looking for any Cells with an `animate-pulse-fast` class. However, this
same class is also used for Cell Highlights. Therefore, previously,
triggering a Cell Highlight would effectively pause the auto-solver for
the duration of the Highlight effect.

To fix: We only consider Cell with the `animate-pulse-fast` class that
don't also have the Cell Highlight class(es).
- Note: We can get away with just detecting one of the classes (the full
  list is just: `bg-slate-300 dark:bg-neutral-500`).