Add bin/conccurrency_test for stress-testing Reveal Actions
This script invokes multiples of the same request at (essentially) the
same time--by spawning multiple threads. This allows for basic
concurrency testing. e.g. for behavior comparison with vs without
pessimistic locks in place.

The current approach is overly simplistic: We find the last {Cell} of
the current {Game}, then we send multiple "Reveal" requests to that
same Cell at(essentially) the same time. Behavior verification is then
left up to the invoker, e.g. by inspecting the Rails server log, or
however.

Must temporarily add this to RevealsController for this script to work:
`skip_before_action :verify_authenticity_token, only: :create`.

Note: This is used to determine behaviors when race conditions arise on
Cell Reveal. The result is as expected... race conditions cause odd
behavior. In the following commit, we'll handle these.