Finish upgrade to Rails 8.0.2 + update gems again
I didn't realize I should've been checking railsdiff.org for changes
even between minor versions of Rails. My bad. So this commit catches up
on the Rails 8.0.1 -> 8.0.2 updates: https://railsdiff.org/8.0.1/8.0.2

Note: We continue to need `libpq-dev` in the Dockerfile under
`# Install packages needed to build gems` in order for postgres to
install properly. (Just verified this again.) So leaving it in there,
but moving it to the end of the list.

Also, we've got a few more gem updates now.