Add an indicator for "Spam" games / Spammers (Users)
When coming across a Game that's been marked as Spam, I'd like to
readily see it. This is really just for me, but not worried if others
notice it too.

Also, add the same indicator for "Spammers" (Users)
- When coming across a User that's played multiples Games that are
  marked as Spam, I'd like to readily see it. This is really just for
  me, but not worried if others notice it too.
  - The threshold for how many "spam" games it takes to be a Spammer is
    currently: 10

Also: Fix to not show spam games when jumping forward/backward via
page nav in Paste Games Index. (We intentionally still do so on the
User-namespaced Prev/Next Games page nav, though.)