Add more indicators to Roster Listings
Add indicators for who created the game and who started the game to
the roster listings for multiplayer games.

As I started on this, I noticed a duplicated concept, hidden in the need
for (what was becoming) 4 arguments to the
Games::Past::ActiveParticipants::Roster::Listing initializer method. The
hidden concept being: the "iteration context". Which would help us to
efficiently reveal which {User} created/started/ended a {Game} without
requiring each iteration to query each of these.
- All that listings now need to care about is the User they wrap, and
  the iteration context. Huzzah!

Also, we take the opportunity to DRY up the view (into shared partials)
and the duplicated portion of the Listing objects (using inheritance).