Use CSS vars for page padding + reference as needed
Now that I've learned about arbitrary CSS variables, this makes so much
sense. Here, we set the variable for page padding (based on breakpoint,
even!) and then reference that as needed:
- For child elements that are all just trying to match the same
  "page"-level padding
- For negative margin offsets of child elements

Also:

Make CSS for horizontal scroll of Game board more robust by referencing
the page padding for its negative margin offsets. Then, there's a lot
of noise as we make this change "universal" to all of the different
Game-related partials and in the different views based on Game state.

Then, finish out utilizing CSS variables in a few more places that I
found that made sense.