Replace simple tables with <dl>s + assoc. utility classes
Especially so for that .dl-collapsible, I didn't like the hacky
structure of the tables. (i.e. making table rows wrap into stacked
columns.)

After recently learning some more tailwindcss-fu, this becomes
easier/cleaner to do with non-tables. In this case, we choose <dl>s
since they're the most semantically correct for a 2-column format, like
we have here.

.dl-basic is a non-collapsible version of .dl-collapsible.

The key to this is the `grid-cols-subgrid` plus the ability to have a
custom-format grid that shrinks down to fit its content, thanks to:
`grid-cols-[minmax(0,max-content)_minmax(0,max-content)]`

Resource:
https://www.youtube.com/watch?v=7d0qmca5kzc