Move User type/classification business logic out of User model
Previously, we defined User#spammer?, but did so wearily. I know we
shouldn't be defining business logic in our ActiveRecord models, but am
still learning how to be better at this. The key is to use service
objects for business logic, of course.

Therefore, this commit moves this logic out of User and into User::Type.
This service includes User-type/classification related predicates:
- dev?
- non_dev?
- spammer?

It's less convenient since you can't be on the command line and just do
`User.first.spammer?`. Instead you must do:
`User::Type.spammer?(User.first)`. But, that's the price we pay for
separation of concerns. While we could just leave a delegate method on
User to allow for `User.first.spammer?`, that's the slippery slope
towards exposing business logic to what we aim to treat simply as our
bag of attributes. So this is a push towards the future we want--with
more to come.
- Note: as far as the consumer of the business logic--the View Model--
  is concerned, this really isn't much different. It's more just "on the
  console" that it feels harder.