Clean up listings iterators
Somehow, an unneeded `each_with_index` made it into my listings
iterators and then got copied around, where it isn't needed. Remove this
in favor of just `each`.

Also, prefer the CSS-based no-listings message in the Dev Portal UI for
Patterns. This makes this the standard across the site and prevents an
extra query on that page.