Don't wrap test files in class-level describe blocks
At some point in the past, class-level describe blocks stopped being
necessary... (not sure when). But, it used to be that if you didn't add
a class-level describe block, test errors would be context-less.

That is, where now we'll see e.g.:

```
 FAIL CellTest::#validate::#coordinates::GIVEN valid, unique #coordinates#test_0001_passes validation (0.24s)
```

... previously we would have only seen:

```
 FAIL #validate::#coordinates::GIVEN valid, unique #coordinates#test_0001_passes validation (0.24s)
```

^ Completely devoid of context as to what Class has failed!

Anyway, now that we see the test name in the failure message, we can get
rid of the top-most level describe blocks.