Refactor `context "GIVEN ` -> `given "` in tests
Since we're already adding `context` as a custom method (via the gemwork
gem), we might as well add an even more direct DSL method.

`given` will be found in most of my test `describe` blocks, as I always
try to setup conditions outside of the `it` blocks.

To support this, we add `given` to test_helper.rb and just have it
delegate to minitest's `describe` method.