Add Wrap and KeywordWrap service objects
These new objects are meant to replace the WrapMethodBehaviors mix-in.

This came about because I've really been wanting to be able to use
keyword arguments on wrapper objects. But the existing
WrapMethodBehaviors version of `wrap` falls short in this regard. With
it, all objects collections must be passed a positional arguments.

Also, lately, I've come to feel that it's better to just straight-up
call an object to perform some behavior. Vs: mixing behavior into an
object--thus, extending it's surface area--and then doing the behavior
within that.

So the combination of these 2 things has led me to want to build a
Wrap service object that could take on this responsibility without
having to mix it in to anything. As I quickly found out... it isn't good
OO-Design to try to force both of these wants into one object, though.
So, we now have Wrap (for positional "objects") and KeywordWrap (for
keyword "objects").

While this felt right, it also felt more verbose. Which ultimately led
me to the idea that each wrapper object could, optionally, define its
own `wrap` class method that takes care of injecting its own settings
into the Wrap or KeywordWrap object and calling it based its passed in
`objects` + any kwargs. This feels like the right balance, at last!

With this, we deprecate the old WrapMethodBehaviors -> `wrap` method. To
be replaced in future commits.