Rename `Tally` -> `#` to save space (esp on mobile)
Previously, the daily tallies on the Games Index page were running off
the right side of the view a bit. Instead of wrapping, I decided to
with a tally symbol (#). And instead of making this conditionally
display only for mobile breakpoints, I figured it's actually more
visually pleasing like this even in desktop.
- But I did add a browser tooltip since this does leave a bit of
  interpretation up to the viewer.