Extract "Bestable Types" from Game
This is a modest start on an effort to extract business logic from
ActiveRecord models. I'm trying to see if I can achieve a system where
the ActiveRecord models are more or less just bags of attributes /
database interfaces, and where all business logic is performed in the
service objects layer.

Game::Type names one of the hidden concepts, around which some basic
functionality exists. It is a similar abstraction to User::Type, which
we also extracted recently. And, so far I'm thinking this may end up
being a pretty common need for ActiveRecord models. Basically, a service
object that deals with "type"-related things in lieu of the ActiveRecord
model. Or... with things that relate to what it means to be that object
type vs any other object type. Or... "is_a?" type concerns.