Keep header aligned to Game board on all breakpoints
Not sure why I ever wanted to let the header elements snap to the edge
of the screen within a certain breakpoint (for xs and sm)... so this
could be a regression, I guess. But, it looks better to keep the header
aligned to the Game board across all breakpoints, from what I can see
right now.