Show past Games timestamp as T+HH:MM:SS if ended today
Previously, we'd just always show the time range like
`<start time stamp> - <end time stamp>` when what I really wanted to
know is how long ago the Game ended. Similar to how we have it now for
just-ended Games.

So now, so long as past Game ended "today" then we'll show a
"T+HH:MM:SS"-style relative timestamp. Older past Games, however, will
still show a start/end timestamp.

The idea is to make it easier to see how long ago someone was last
playing a Game. This makes it easier to know to jump in or not to
hopefully get a multi-player Game going, when you might not otherwise
expect to be able to.

Also, move Games::JustEnded::ElapsedTime -> Games::Past::ElapsedTime
now that both just-ended and past Games need it. Because it makes more
sense to me that the default case here is that this is a "Past" Game
even if that is, specifically, just as-of the past 3 minutes (aka a
just-ended Game).