Rearrange display of Daily Engagement Tally on listings
For listing groups, the display of engagement tallies felt unbalanced
having the sum at the end. Especially due to largely varying position
based on varying content in the Alliance / Mines tallies.

So, we rearrange to show the sum first and it feels more balanced.

Also, change from the `#` text to the #️⃣ emoji as it feels a bit more
distinctive this way.
- I would have just used `Tally: ` but it causes the overall line length
  to not fit on mobile.

To accomplish this, I've simply duplicated the content out of
app/views/games/past/_engagement_tally.html.erb ... I don't feel it's
worth DRYing up from there. We'll just have to live with the duplication
(or do something about it later, when we know more).

Also: Add a bit more space between days (listings grouped by date) in
Games Index page. This makes it easier to parse. And why not?... we have
plenty for vertical room!