Refactor: Extract Cell::Neighbors from Cell
This continues my quest to extract business logic from ActiveRecord
models.

Cell::Neighbors is a value object (immutable and all) that: represents
the collection of neighboring Cells for the given Cell, which is based
on the corresponding Coordinates#neighbors.

This extraction allows us to remove the
`rubocop:disable Metrics/ClassLength` override on Cell 🎉

There are no functional changes herein.