Refactor: Rename Cell-level services -> Board-level services
The reason for this refactor is to make way for Cell-level service
objects to exist. e.g. to allow for Cell::Reveal to do the actual work
of revealing a Cell.

While considering names/namespacing, it became clear to me that it's
better to think of the old Cell::Reveal service as more of a Game
Board-level thing. i.e. the existing services were really about the
interrelationships between the Game/Board and a given Cell. They
reveal *other* Cells within the Board, for example.

Next up, we can now reclaim the Cell-level service objects to perform
the actual Cell-level services (things that aren't based on
associations to other Cells on the board, like revealing an individual
Cell).

Note: With this refactor, Rails now has trouble automatically inferring
the Game <-> Board relationship in the Game `has_one :board`
association. It resolves the association as the new Game::Board modules
instead of the ::Board Model, as we might expect. So we have to help
Rails out and specify the `class_name: "::Board"` option.