Dehighlight Cells on flag
This is a "Just in case" type thing... though I may have seen a race
condition once where it could have helped. i.e. I think I once saw a
flagged AND highlighted Cell. Which makes no sense.

Note: We only dehighlight cells on flag. We don't dehighlight on unflag.
I could be wrong but this seems to make sense right now. Change this if
there's ever proof to the contrary... just seems like since a highlight
is valid on an unflagged Cell we shouldn't negate this. However, on a
flagged Cell, a highlight just doesn't make sense--ever.