Remove Cell#neighbors i/f/o Cell::Neighbors
This continues the ongoing effort to extract business logic out of the
ActiveRecord models and into separate service objects, etc.

The previous commits were all marching towards this point, where we
could remove Cell#neighbors and just rely fully on Cell::Neighbors
instead.

To support Game::Board::RecursiveReveal, we also implement
Cell::Neighbors#unrevealed_cells

Of note, we also now get to remove the reek overrides in Cell:
- TooManyMethods
- RepeatedConditional
... we may have been able to do this a commit or 2 ago, in reality, but
I missed it if so. And this is a good time to do it.