Tidy: Freeze value objects
If what we really have are value objects, then they should be immutable.

I wasn't able to apply this to Duration, though, since it uses internal
caching and I didn't want to take that away from it... I do think it's a
value object still, but... kicking that can down the road.