End use of `unit_class` in tests
Instead of abstracting the class name for a test file into a
`unit_class` method (let block), we'll now just repeat the class name
throughout the test file.

The goal is to be more expressive. Ways in which this is more
expressive:
1. Having to know/remember that there is a `unit_class` method is a
   (small) hurdle
2. Seeing `unit_class` within a test example introduces indirection as
   one has to figure out what that is in the context.
   - And even more so when testing a subclass within a main test file.
3. constant/namespace look-ups that include `unit_class` are odd. e.g.
   `subject { unit_class::Settings }` vs `subject { Board::Settings }`.

Also, in a few files, I've distributed the `subject` out, into each of
the `describe` blocks, as is my latest style. Again, to be more
expressive... in the sense that we want to be able to see the subject
of the test close to the test vs at the top of the file.