Refactor "wrapper_class" to TestDouble in Wrap tests
This is a continuation of my effort to make tests more expressive. I
think the plain class names are easier to understand (straight away)
than the `let(:wrapper_class)` blocks.

Also, I prefer subclasses to be defined at the end of the file vs the
top. After switching to class definitions, this is feels more obvious.