Extract Game factory methods out to Game::Factory service
This continues the ongoing effort to extract business logic out of the
ActiveRecord models and into separate service objects, etc.

I'm not sure if this refactoring is the proper move... but trying it
out. I like that it extracts the responsibility from the ActiveRecord
model... which already has its own ways of building objects. Plus, I
like the explicit naming and separated-out tests for these somewhat
unique methods.