Add Board::Settings#dimensions
I remember not wanting to add this in the past, but adding now because
I'm trying to make extracting the settings-related methods from Board
easy (if I even decide to go through with it). Either way, though, I
think it makes enough sense for Board::Settings to own this instead of
Board at this point.