Add Conversions.CoordinatesArray
Since we already have the CoordinatesArray object, we should use it!
e.g. Instead of creating an Array object that hold a bunch of
Coordinates objects.

Also, now that we have this conversion, we can use it for object
comparison in CoordinatesArray#<=> (the same as we just did in the
previous commit for Coordinates).