Rescue + refresh on "Game Over" DispatchAction exceptions
Previously, we were just raising a
Games::Current::Board::Cells::DispatchAction::Error exception if a
DispatchAction was attempted after the Game had already ended. This was
fine for a while, but now it's established behavior and I get exception
reports for each one. They're not too frequent, but none of them are
needed. This is just a race condition that can occur any time 2+ players
are dispatching actions ("clicking", generally) right around the same
time as a Game-ending move occurred.

To fix, change the exception type to
Games::Current::Board::Cells::DispatchAction::GameOverError, which is
more descriptive. This gives me the confidence to rescue on that error
and know what action to take in response to it, within each of the
Action controllers.