Rescue + refresh on "Ready Only" DispatchEffect exceptions
Previously, we were just letting ActiveRecord raise a
ActiveRecord::ReadOnlyRecord exception if a
DispatchEffect was attempted after the Game had already ended. This was
fine for a while, but now it's established behavior and I get exception
reports for each one. They're not too frequent, but none of them are
needed. This is just a race condition that can occur any time 2+ players
are do highlight effects right around the same time as a Game-ending
move has occurred.

To fix, rescue on ActiveRecord::ReadOnlyRecord within DispatchEffect,
translate that into a
Games::Current::Board::Cells::DispatchEffect::GameOverError error for
clearer context, then rescue on *that* within the controller and refresh
the page.
- This "adapter"/translation on the error type also allows us to change
  how we recognize and respond to a "Game Over" event within
  DispatchEffect without having to change the controller actions.